+++ opts/swarmopts/port_test.go
@@ -1,4 +1,4 @@
-package opts
+package swarmopts
 
 import (
 	"bytes"
